Move ast node classes to their own file Allows a star import by users of the ast nodes without catching anything else
